Fixed validation long time for error widgets
I have made a verification, to check if the class has in valid value then compare the error message, this way we don't have to search for an element that may no appear at all and waste a long time to wait.
I have tested this on e-mail field, the test took 3 minutes, and after fixing this issue, it took only 29 seconds.